Temperature field change to decimalfield

+++ visitor/models.py
@@ -40,7 +40,7 @@ class RequestForm(models.Model):
     infection = models.BooleanField(editable=True)
     typhoid = models.BooleanField(editable=True)
     country = models.CharField(max_length=100, verbose_name="Last Countries Visited")
-    temp = models.IntegerField()
+    temp = models.DecimalField(max_digits=2, decimal_places=2)
 
     agreement = models.BooleanField(editable=True, default=True)
     dateSigned = models.DateField(auto_now=False, verbose_name='Date Signed')
